test: fold issue94 test file into handler_test.go by resource
Delete handler_events_issue94_test.go — naming tests after the ticket that
spawned them is an anti-pattern; tests belong with the resource they cover.

- Move the shared request-body builders (negRiskEventSingleMarketBody,
  addBinaryMarketBody, addNegRiskMarketBody) into handler_test.go next to
  binaryEventBody. handler_test.go already depended on these, so the
  dependency direction was backwards.
- Drop binaryEventSingleMarketBody (a duplicate of binaryEventBody) and the
  four redundant issue94 tests already covered by the create/AddMarkets
  sections (Binary singular-payload, Binary append, NegRisk append uses
  stored market id, NegRisk append wrong-parent rejection).
- Fold in the one piece of unique coverage: a genuine NegRisk create
  happy-path. The existing _Success name was on a multi-market rejection
  test, so rename that to _MultiMarketRejected (and the binary equivalent to
  _Binary_MultiMarketRejected) and reclaim _Success for the real success case.
